refactor: split the script into modules
