+++ task-02-object-gauss/src/scripts/main.js
@@ -1,0 +1,3 @@
+import { runGenerate } from './gaussLogic.js';
+
+runGenerate();
