+++ task-02-object-gauss/src/scripts/styles.js
@@ -1,0 +1,15 @@
+import { INPUT_NUMBER } from './utils.js';
+
+export function useDefaultColors() {
+  INPUT_NUMBER.classList.remove('input-focus-colors');
+  INPUT_NUMBER.classList.remove('input-error-colors');
+}
+
+export function useFocusColors() {
+  INPUT_NUMBER.classList.add('input-focus-colors');
+  INPUT_NUMBER.classList.remove('input-error-colors');
+}
+
+export function useErrorColors() {
+  INPUT_NUMBER.classList.add('input-error-colors');
+}
